Modify constant K8sNetworkPolicyRulePriority's type to int32 (#189)
* modify constant K8sNetworkPolicyRulePriority to int32

* add correlate fields support

* add tests

+++ pkg/registry/registry.go
@@ -55,7 +55,7 @@ const (
 
 // placeholder of NetworkPolicyRulePriority for K8s Network Policy.
 const (
-	K8sNetworkPolicyRulePriority = int16(-1)
+	K8sNetworkPolicyRulePriority = int32(-1)
 )
 
 // enum for flowEndReason field in IANA registry.
